add 'merge' action
