+++ action.yml
@@ -58,4 +58,4 @@ outputs:
       Common uses cases for such a download URL can be adding download links to artifacts in descriptions or comments on pull requests or issues.
 runs:
   using: 'node20'
-  main: 'dist/index.js'
+  main: 'dist/upload/index.js'
