+++ package.json
@@ -2,10 +2,10 @@
   "name": "upload-artifact",
   "version": "4.2.0",
   "description": "Upload an Actions Artifact in a workflow run",
-  "main": "dist/index.js",
+  "main": "dist/upload/index.js",
   "scripts": {
     "build": "tsc",
-    "release": "ncc build src/upload-artifact.ts && git add -f dist/index.js",
+    "release": "ncc build src/upload/index.ts -o dist/upload && ncc build src/merge/index.ts -o dist/merge",
     "check-all": "concurrently \"npm:format-check\" \"npm:lint\" \"npm:test\" \"npm:build\"",
     "format": "prettier --write **/*.ts",
     "format-check": "prettier --check **/*.ts",
@@ -33,7 +33,8 @@
     "@actions/core": "^1.10.0",
     "@actions/github": "^6.0.0",
     "@actions/glob": "^0.3.0",
-    "@actions/io": "^1.1.2"
+    "@actions/io": "^1.1.2",
+    "minimatch": "^9.0.3"
   },
   "devDependencies": {
     "@types/jest": "^29.2.5",
