+++ src/merge/constants.ts
@@ -1,0 +1,9 @@
+/* eslint-disable no-unused-vars */
+export enum Inputs {
+  Into = 'into',
+  Pattern = 'pattern',
+  SeparateDirectories = 'separate-directories',
+  RetentionDays = 'retention-days',
+  CompressionLevel = 'compression-level',
+  DeleteMerged = 'delete-merged'
+}
